SynthBenchmark: Fixed conversion warnings. Enabled UnsafeTypeCast warnings as errors.
#rb Catalin.Dragoiu

[CL 28340939 by ionut matasaru in ue5-main branch]

+++ Engine/Source/Runtime/SynthBenchmark/Private/SynthBenchmarkPrivate.cpp
@@ -63,7 +63,7 @@ static FTimeSample RunBenchmark(float WorkScale, float (*Function)())
 		FPlatformMisc::MemoryBarrier();
 	}
 	
-	return FTimeSample(Sum, Sum / RunCount);
+	return FTimeSample(Sum, Sum / (float)RunCount);
 }
 
 template<int NumMethods>
